# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TalentLock API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /job-requirements:
    get:
      operationId: listJobRequirements
      tags: [job-requirements]
      summary: List job requirements
      parameters:
        - name: employerId
          in: query
          schema:
            type: integer
          description: Filter by employer
        - name: status
          in: query
          schema:
            type: string
          description: Filter by status (open, filled, closed)
      responses:
        "200":
          description: List of job requirements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JobRequirement"
    post:
      operationId: createJobRequirement
      tags: [job-requirements]
      summary: Create a job requirement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobRequirementBody"
      responses:
        "201":
          description: Created job requirement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobRequirement"
  /job-requirements/{id}:
    get:
      operationId: getJobRequirement
      tags: [job-requirements]
      summary: Get a job requirement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job requirement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobRequirement"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    patch:
      operationId: updateJobRequirement
      tags: [job-requirements]
      summary: Update a job requirement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateJobRequirementBody"
      responses:
        "200":
          description: Updated job requirement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobRequirement"
    delete:
      operationId: deleteJobRequirement
      tags: [job-requirements]
      summary: Delete a job requirement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ── User ──────────────────────────────────────────────────────────────────
...
    ProfessionCategory:
      type: string
      enum: [technology, education]
...
    RateType:
      type: string
      enum: [hourly, per_day, per_session, per_course]
...
    JobRequirement:
      type: object
      properties:
        id:
          type: integer
        employerId:
          type: integer
        employerVerificationLevel:
          type: string
          enum: [unverified, partially_verified, fully_verified]
        title:
          type: string
        fieldOfWork:
          type: string
        description:
          type: string
        requiredSkills:
          type: array
          items:
            type: string
        minExperience:
          type: integer
        paymentType:
          type: string
          description: "hourly or daily"
        budget:
          type: ["number", "null"]
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        status:
          type: string
          description: "open, filled, closed"
        professionCategory:
          $ref: "#/components/schemas/ProfessionCategory"
        rateType:
          $ref: "#/components/schemas/RateType"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - employerId
        - employerVerificationLevel
        - title
        - fieldOfWork
        - description
        - requiredSkills
        - minExperience
        - paymentType
        - startDate
        - endDate
        - status
        - professionCategory
        - rateType
        - createdAt
    CreateJobRequirementBody:
      type: object
      properties:
        title:
          type: string
        fieldOfWork:
          type: string
        description:
          type: string
        requiredSkills:
          type: array
          items:
            type: string
        minExperience:
          type: integer
        paymentType:
          type: string
        budget:
          type: ["number", "null"]
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        professionCategory:
          $ref: "#/components/schemas/ProfessionCategory"
        rateType:
          $ref: "#/components/schemas/RateType"
      required:
        - title
        - fieldOfWork
        - description
        - requiredSkills
        - minExperience
        - paymentType
        - startDate
        - endDate
    UpdateJobRequirementBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        requiredSkills:
          type: array
          items:
            type: string
        minExperience:
          type: integer
        budget:
          type: ["number", "null"]
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        status:
          type: string
        professionCategory:
          $ref: "#/components/schemas/ProfessionCategory"
        rateType:
          $ref: "#/components/schemas/RateType"

    # ── Booking ────────────────────────────────────────────────────────────────
...