# lib/api-zod/src/generated/types/jobRequirement.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TalentLock API specification
 * OpenAPI spec version: 0.1.0
 */
import type { JobRequirementEmployerVerificationLevel } from "./jobRequirementEmployerVerificationLevel";
import type { ProfessionCategory } from "./professionCategory";
import type { RateType } from "./rateType";

export interface JobRequirement {
  id: number;
  employerId: number;
  employerVerificationLevel: JobRequirementEmployerVerificationLevel;
  title: string;
  fieldOfWork: string;
  description: string;
  requiredSkills: string[];
  minExperience: number;
  /** hourly or daily */
  paymentType: string;
  /** @nullable */
  budget?: number | null;
  startDate: Date;
  endDate: Date;
  /** open, filled, closed */
  status: string;
  professionCategory: ProfessionCategory;
  rateType: RateType;
  createdAt: Date;
}